Move ReactNativeFile class to index file
