--- src/ReactNativeFile.ts
@@ -1,11 +1,0 @@
-export class ReactNativeFile {
-    private uri: string;
-    private type: string;
-    private name: string;
-
-    constructor(params: { uri: string; type: string; name: string }) {
-        this.uri = params.uri;
-        this.type = params.type;
-        this.name = params.name;
-    }
-}
