+++ tslint.json
@@ -7,7 +7,8 @@
   "jsRules": {},
   "rules": {
     "ordered-imports": true,
-    "prettier": true
+    "prettier": true,
+    "max-classes-per-file": false
   },
   "rulesDirectory": ["tslint-plugin-prettier"]
 }
